Bruke mer idiomatisk action pattern
